+++ x/concentrated-liquidity/swaps_test.go
@@ -32,7 +32,7 @@ type SwapTest struct {
 	tokenInDenom string
 
 	// Shared.
-	priceLimit               osmomath.Dec
+	priceLimit               osmomath.BigDec
 	spreadFactor             osmomath.Dec
 	secondPositionLowerPrice osmomath.Dec
 	secondPositionUpperPrice osmomath.Dec
@@ -84,7 +84,7 @@ var (
 		"single position within one tick: usdc -> eth": {
 			tokenIn:       sdk.NewCoin("usdc", osmomath.NewInt(42000000)),
 			tokenOutDenom: "eth",
-			priceLimit:    osmomath.NewDec(5004),
+			priceLimit:    osmomath.NewBigDec(5004),
 			spreadFactor:  osmomath.ZeroDec(),
 			// from math import *
 			// from decimal import *
@@ -125,7 +125,7 @@ var (
 		"single position within one tick: usdc -> eth, with zero price limit": {
 			tokenIn:       sdk.NewCoin("usdc", osmomath.NewInt(42000000)),
 			tokenOutDenom: "eth",
-			priceLimit:    osmomath.ZeroDec(),
+			priceLimit:    osmomath.ZeroBigDec(),
 			spreadFactor:  osmomath.ZeroDec(),
 			// from math import *
 			// from decimal import *
@@ -166,7 +166,7 @@ var (
 		"single position within one tick: eth -> usdc": {
 			tokenIn:       sdk.NewCoin("eth", osmomath.NewInt(13370)),
 			tokenOutDenom: "usdc",
-			priceLimit:    osmomath.NewDec(4993),
+			priceLimit:    osmomath.NewBigDec(4993),
 			spreadFactor:  osmomath.ZeroDec(),
 			// from math import *
 			// from decimal import *
@@ -203,7 +203,7 @@ var (
 		"single position within one tick: eth -> usdc, with zero price limit": {
 			tokenIn:       sdk.NewCoin("eth", osmomath.NewInt(13370)),
 			tokenOutDenom: "usdc",
-			priceLimit:    osmomath.ZeroDec(),
+			priceLimit:    osmomath.ZeroBigDec(),
 			spreadFactor:  osmomath.ZeroDec(),
 			// from math import *
 			// from decimal import *
@@ -241,7 +241,7 @@ var (
 		"two positions within one tick: usdc -> eth": {
 			tokenIn:                  sdk.NewCoin("usdc", osmomath.NewInt(42000000)),
 			tokenOutDenom:            "eth",
-			priceLimit:               osmomath.NewDec(5002),
+			priceLimit:               osmomath.NewBigDec(5002),
 			spreadFactor:             osmomath.ZeroDec(),
 			secondPositionLowerPrice: DefaultLowerPrice,
 			secondPositionUpperPrice: DefaultUpperPrice,
@@ -277,7 +277,7 @@ var (
 		"two positions within one tick: eth -> usdc": {
 			tokenIn:                  sdk.NewCoin("eth", osmomath.NewInt(13370)),
 			tokenOutDenom:            "usdc",
-			priceLimit:               osmomath.NewDec(4996),
+			priceLimit:               osmomath.NewBigDec(4996),
 			spreadFactor:             osmomath.ZeroDec(),
 			secondPositionLowerPrice: DefaultLowerPrice,
 			secondPositionUpperPrice: DefaultUpperPrice,
@@ -321,7 +321,7 @@ var (
 		"two positions with consecutive price ranges: usdc -> eth": {
 			tokenIn:                  sdk.NewCoin("usdc", osmomath.NewInt(10000000000)),
 			tokenOutDenom:            "eth",
-			priceLimit:               osmomath.NewDec(6255),
+			priceLimit:               osmomath.NewBigDec(6255),
 			spreadFactor:             osmomath.ZeroDec(),
 			secondPositionLowerPrice: osmomath.NewDec(5500),
 			secondPositionUpperPrice: osmomath.NewDec(6250),
@@ -376,7 +376,7 @@ var (
 		"two positions with consecutive price ranges: eth -> usdc": {
 			tokenIn:                  sdk.NewCoin("eth", osmomath.NewInt(2000000)),
 			tokenOutDenom:            "usdc",
-			priceLimit:               osmomath.NewDec(3900),
+			priceLimit:               osmomath.NewBigDec(3900),
 			spreadFactor:             osmomath.ZeroDec(),
 			secondPositionLowerPrice: osmomath.NewDec(4000),
 			secondPositionUpperPrice: osmomath.NewDec(4545),
@@ -433,7 +433,7 @@ var (
 		"two positions with partially overlapping price ranges: usdc -> eth": {
 			tokenIn:                  sdk.NewCoin("usdc", osmomath.NewInt(10000000000)),
 			tokenOutDenom:            "eth",
-			priceLimit:               osmomath.NewDec(6056),
+			priceLimit:               osmomath.NewBigDec(6056),
 			spreadFactor:             osmomath.ZeroDec(),
 			secondPositionLowerPrice: osmomath.NewDec(5001),
 			secondPositionUpperPrice: osmomath.NewDec(6250),
@@ -487,7 +487,7 @@ var (
 		"two positions with partially overlapping price ranges, not utilizing full liquidity of second position: usdc -> eth": {
 			tokenIn:       sdk.NewCoin("usdc", osmomath.NewInt(8500000000)),
 			tokenOutDenom: "eth",
-			priceLimit:    osmomath.NewDec(6056),
+			priceLimit:    osmomath.NewBigDec(6056),
 			spreadFactor:  osmomath.ZeroDec(),
 			// from math import *
 			// from decimal import *
@@ -553,7 +553,7 @@ var (
 		"two positions with partially overlapping price ranges: eth -> usdc": {
 			tokenIn:       sdk.NewCoin("eth", osmomath.NewInt(2000000)),
 			tokenOutDenom: "usdc",
-			priceLimit:    osmomath.NewDec(4128),
+			priceLimit:    osmomath.NewBigDec(4128),
 			spreadFactor:  osmomath.ZeroDec(),
 			// from math import *
 			// from decimal import *
@@ -617,7 +617,7 @@ var (
 		"two positions with partially overlapping price ranges, not utilizing full liquidity of second position: eth -> usdc": {
 			tokenIn:                  sdk.NewCoin("eth", osmomath.NewInt(1800000)),
 			tokenOutDenom:            "usdc",
-			priceLimit:               osmomath.NewDec(4128),
+			priceLimit:               osmomath.NewBigDec(4128),
 			spreadFactor:             osmomath.ZeroDec(),
 			secondPositionLowerPrice: osmomath.NewDec(4000),
 			secondPositionUpperPrice: osmomath.NewDec(4999),
@@ -678,7 +678,7 @@ var (
 		"two sequential positions with a gap": {
 			tokenIn:                  sdk.NewCoin("usdc", osmomath.NewInt(10000000000)),
 			tokenOutDenom:            "eth",
-			priceLimit:               osmomath.NewDec(6106),
+			priceLimit:               osmomath.NewBigDec(6106),
 			spreadFactor:             osmomath.ZeroDec(),
 			secondPositionLowerPrice: osmomath.NewDec(5501),
 			secondPositionUpperPrice: osmomath.NewDec(6250),
@@ -734,7 +734,7 @@ var (
 		"single position within one tick, trade completes but slippage protection interrupts trade early: eth -> usdc": {
 			tokenIn:       sdk.NewCoin("eth", osmomath.NewInt(13370)),
 			tokenOutDenom: "usdc",
-			priceLimit:    osmomath.NewDec(4994),
+			priceLimit:    osmomath.NewBigDec(4994),
 			spreadFactor:  osmomath.ZeroDec(),
 			// from math import *
 			// from decimal import *
@@ -774,7 +774,7 @@ var (
 			// are estimated by utilizing scripts from scripts/cl/main.py
 			tokenIn:           sdk.NewCoin("usdc", osmomath.NewInt(42000000)),
 			tokenOutDenom:     "eth",
-			priceLimit:        osmomath.NewDec(5004),
+			priceLimit:        osmomath.NewBigDec(5004),
 			spreadFactor:      osmomath.MustNewDecFromStr("0.01"),
 			expectedTokenIn:   sdk.NewCoin("usdc", osmomath.NewInt(42000000)),
 			expectedTokenOut:  sdk.NewCoin("eth", osmomath.NewInt(8312)),
@@ -790,7 +790,7 @@ var (
 			// are estimated by utilizing scripts from scripts/cl/main.py
 			tokenIn:                  sdk.NewCoin("eth", osmomath.NewInt(13370)),
 			tokenOutDenom:            "usdc",
-			priceLimit:               osmomath.NewDec(4990),
+			priceLimit:               osmomath.NewBigDec(4990),
 			spreadFactor:             osmomath.MustNewDecFromStr("0.03"),
 			secondPositionLowerPrice: DefaultLowerPrice,
 			secondPositionUpperPrice: DefaultUpperPrice,
@@ -811,7 +811,7 @@ var (
 			// are estimated by utilizing scripts from scripts/cl/main.py
 			tokenIn:                  sdk.NewCoin("eth", osmomath.NewInt(2000000)),
 			tokenOutDenom:            "usdc",
-			priceLimit:               osmomath.NewDec(4094),
+			priceLimit:               osmomath.NewBigDec(4094),
 			spreadFactor:             osmomath.MustNewDecFromStr("0.05"),
 			secondPositionLowerPrice: osmomath.NewDec(4000),
 			secondPositionUpperPrice: osmomath.NewDec(4545),
@@ -831,7 +831,7 @@ var (
 			// are estimated by utilizing scripts from scripts/cl/main.py
 			tokenIn:                  sdk.NewCoin("usdc", osmomath.NewInt(10000000000)),
 			tokenOutDenom:            "eth",
-			priceLimit:               osmomath.NewDec(6056),
+			priceLimit:               osmomath.NewBigDec(6056),
 			spreadFactor:             osmomath.MustNewDecFromStr("0.1"),
 			secondPositionLowerPrice: osmomath.NewDec(5001),
 			secondPositionUpperPrice: osmomath.NewDec(6250),
@@ -851,7 +851,7 @@ var (
 			// are estimated by utilizing scripts from scripts/cl/main.py
 			tokenIn:                  sdk.NewCoin("eth", osmomath.NewInt(1800000)),
 			tokenOutDenom:            "usdc",
-			priceLimit:               osmomath.NewDec(4128),
+			priceLimit:               osmomath.NewBigDec(4128),
 			spreadFactor:             osmomath.MustNewDecFromStr("0.005"),
 			secondPositionLowerPrice: osmomath.NewDec(4000),
 			secondPositionUpperPrice: osmomath.NewDec(4999),
@@ -871,7 +871,7 @@ var (
 			// are estimated by utilizing scripts from scripts/cl/main.py
 			tokenIn:                  sdk.NewCoin("usdc", osmomath.NewInt(10000000000)),
 			tokenOutDenom:            "eth",
-			priceLimit:               osmomath.NewDec(6106),
+			priceLimit:               osmomath.NewBigDec(6106),
 			secondPositionLowerPrice: osmomath.NewDec(5501),
 			secondPositionUpperPrice: osmomath.NewDec(6250),
 			spreadFactor:             osmomath.MustNewDecFromStr("0.03"),
@@ -890,7 +890,7 @@ var (
 			// are estimated by utilizing scripts from scripts/cl/main.py
 			tokenIn:          sdk.NewCoin("eth", osmomath.NewInt(13370)),
 			tokenOutDenom:    "usdc",
-			priceLimit:       osmomath.NewDec(4994),
+			priceLimit:       osmomath.NewBigDec(4994),
 			spreadFactor:     osmomath.MustNewDecFromStr("0.01"),
 			expectedTokenIn:  sdk.NewCoin("eth", osmomath.NewInt(13023)),
 			expectedTokenOut: sdk.NewCoin("usdc", osmomath.NewInt(64417624)),
@@ -907,14 +907,14 @@ var (
 		"single position within one tick, trade does not complete due to lack of liquidity: usdc -> eth": {
 			tokenIn:       sdk.NewCoin("usdc", osmomath.NewInt(5300000000)),
 			tokenOutDenom: "eth",
-			priceLimit:    osmomath.NewDec(6000),
+			priceLimit:    osmomath.NewBigDec(6000),
 			spreadFactor:  osmomath.ZeroDec(),
 			expectErr:     true,
 		},
 		"single position within one tick, trade does not complete due to lack of liquidity: eth -> usdc": {
 			tokenIn:       sdk.NewCoin("eth", osmomath.NewInt(1100000)),
 			tokenOutDenom: "usdc",
-			priceLimit:    osmomath.NewDec(4000),
+			priceLimit:    osmomath.NewBigDec(4000),
 			spreadFactor:  osmomath.ZeroDec(),
 			expectErr:     true,
 		},
@@ -929,7 +929,7 @@ var (
 		"single position within one tick: eth (in) -> usdc (out) | zfo": {
 			tokenOut:     sdk.NewCoin(USDC, osmomath.NewInt(42000000)),
 			tokenInDenom: ETH,
-			priceLimit:   osmomath.NewDec(4993),
+			priceLimit:   osmomath.NewBigDec(4993),
 			spreadFactor: osmomath.ZeroDec(),
 			// from math import *
 			// from decimal import *
@@ -958,7 +958,7 @@ var (
 		"single position within one tick: usdc (in) -> eth (out) ofz": {
 			tokenOut:     sdk.NewCoin(ETH, osmomath.NewInt(13370)),
 			tokenInDenom: USDC,
-			priceLimit:   osmomath.NewDec(5010),
+			priceLimit:   osmomath.NewBigDec(5010),
 			spreadFactor: osmomath.ZeroDec(),
 			// from math import *
 			// from decimal import *
@@ -987,7 +987,7 @@ var (
 		"two positions within one tick: eth (in) -> usdc (out) | zfo": {
 			tokenOut:                 sdk.NewCoin("usdc", osmomath.NewInt(66829187)),
 			tokenInDenom:             "eth",
-			priceLimit:               osmomath.NewDec(4990),
+			priceLimit:               osmomath.NewBigDec(4990),
 			spreadFactor:             osmomath.ZeroDec(),
 			secondPositionLowerPrice: DefaultLowerPrice,
 			secondPositionUpperPrice: DefaultUpperPrice,
@@ -1024,7 +1024,7 @@ var (
 		"two positions within one tick: usdc (in) -> eth (out) | ofz": {
 			tokenOut:                 sdk.NewCoin("eth", osmomath.NewInt(8398)),
 			tokenInDenom:             "usdc",
-			priceLimit:               osmomath.NewDec(5020),
+			priceLimit:               osmomath.NewBigDec(5020),
 			spreadFactor:             osmomath.ZeroDec(),
 			secondPositionLowerPrice: DefaultLowerPrice,
 			secondPositionUpperPrice: DefaultUpperPrice,
@@ -1058,7 +1058,7 @@ var (
 		"two positions with consecutive price ranges: eth (in) -> usdc (out) | zfo": {
 			tokenOut:                 sdk.NewCoin("usdc", osmomath.NewInt(9103422788)),
 			tokenInDenom:             "eth",
-			priceLimit:               osmomath.NewDec(3900),
+			priceLimit:               osmomath.NewBigDec(3900),
 			spreadFactor:             osmomath.ZeroDec(),
 			secondPositionLowerPrice: osmomath.NewDec(4000),
 			secondPositionUpperPrice: osmomath.NewDec(4545),
@@ -1119,7 +1119,7 @@ var (
 		"two positions with consecutive price ranges: usdc (in) -> eth (out) | ofz": {
 			tokenOut:                 sdk.NewCoin(ETH, osmomath.NewInt(1820630)),
 			tokenInDenom:             USDC,
-			priceLimit:               osmomath.NewDec(6106),
+			priceLimit:               osmomath.NewBigDec(6106),
 			spreadFactor:             osmomath.ZeroDec(),
 			secondPositionLowerPrice: osmomath.NewDec(5500), // 315000
 			secondPositionUpperPrice: osmomath.NewDec(6250), // 322500
@@ -1177,7 +1177,7 @@ var (
 		"two positions with partially overlapping price ranges: eth (in) -> usdc (out) | zfo": {
 			tokenOut:                 sdk.NewCoin(USDC, osmomath.NewInt(9321276930)),
 			tokenInDenom:             ETH,
-			priceLimit:               osmomath.NewDec(4128),
+			priceLimit:               osmomath.NewBigDec(4128),
 			spreadFactor:             osmomath.ZeroDec(),
 			secondPositionLowerPrice: osmomath.NewDec(4000),
 			secondPositionUpperPrice: osmomath.NewDec(4999),
@@ -1246,7 +1246,7 @@ var (
 		"two positions with partially overlapping price ranges, not utilizing full liquidity of second position: eth (in) -> usdc (out) | zfo": {
 			tokenOut:                 sdk.NewCoin(USDC, osmomath.NewInt(8479320318)),
 			tokenInDenom:             ETH,
-			priceLimit:               osmomath.NewDec(4128),
+			priceLimit:               osmomath.NewBigDec(4128),
 			spreadFactor:             osmomath.ZeroDec(),
 			secondPositionLowerPrice: osmomath.NewDec(4000),
 			secondPositionUpperPrice: osmomath.NewDec(4999),
@@ -1317,7 +1317,7 @@ var (
 		"two positions with partially overlapping price ranges: usdc (in) -> eth (out) | ofz": {
 			tokenOut:                 sdk.NewCoin(ETH, osmomath.NewInt(1864161)),
 			tokenInDenom:             USDC,
-			priceLimit:               osmomath.NewDec(6056),
+			priceLimit:               osmomath.NewBigDec(6056),
 			spreadFactor:             osmomath.ZeroDec(),
 			secondPositionLowerPrice: osmomath.NewDec(5001),
 			secondPositionUpperPrice: osmomath.NewDec(6250),
@@ -1378,7 +1378,7 @@ var (
 		"two positions with partially overlapping price ranges, not utilizing full liquidity of second position: usdc (in) -> eth (out) | ofz": {
 			tokenOut:                 sdk.NewCoin(ETH, osmomath.NewInt(1609138)),
 			tokenInDenom:             USDC,
-			priceLimit:               osmomath.NewDec(6056),
+			priceLimit:               osmomath.NewBigDec(6056),
 			spreadFactor:             osmomath.ZeroDec(),
 			secondPositionLowerPrice: osmomath.NewDec(5001),
 			secondPositionUpperPrice: osmomath.NewDec(6250),
@@ -1451,7 +1451,7 @@ var (
 		"two sequential positions with a gap usdc (in) -> eth (out) | ofz": {
 			tokenOut:                 sdk.NewCoin(ETH, osmomath.NewInt(1820545)),
 			tokenInDenom:             USDC,
-			priceLimit:               osmomath.NewDec(6106),
+			priceLimit:               osmomath.NewBigDec(6106),
 			spreadFactor:             osmomath.ZeroDec(),
 			secondPositionLowerPrice: osmomath.NewDec(5501), // 315010
 			secondPositionUpperPrice: osmomath.NewDec(6250), // 322500
@@ -1504,7 +1504,7 @@ var (
 		"single position within one tick, trade completes but slippage protection interrupts trade early: usdc (in) -> eth (out) | ofz": {
 			tokenOut:     sdk.NewCoin(ETH, osmomath.NewInt(1820545)),
 			tokenInDenom: USDC,
-			priceLimit:   osmomath.NewDec(5002),
+			priceLimit:   osmomath.NewBigDec(5002),
 			spreadFactor: osmomath.ZeroDec(),
 			// from math import *
 			// from decimal import *
@@ -1532,7 +1532,7 @@ var (
 		"spread factor 1: single position within one tick: eth (in) -> usdc (out) (1% spread factor) | zfo": {
 			tokenOut:     sdk.NewCoin(USDC, osmomath.NewInt(42000000)),
 			tokenInDenom: ETH,
-			priceLimit:   osmomath.NewDec(4993),
+			priceLimit:   osmomath.NewBigDec(4993),
 			spreadFactor: osmomath.MustNewDecFromStr("0.01"),
 			// from math import *
 			// from decimal import *
@@ -1569,7 +1569,7 @@ var (
 		"spread factor 2: two positions within one tick: usdc (in) -> eth (out) (3% spread factor) | ofz": {
 			tokenOut:                 sdk.NewCoin(ETH, osmomath.NewInt(8398)),
 			tokenInDenom:             USDC,
-			priceLimit:               osmomath.NewDec(5020),
+			priceLimit:               osmomath.NewBigDec(5020),
 			spreadFactor:             osmomath.MustNewDecFromStr("0.03"),
 			secondPositionLowerPrice: DefaultLowerPrice,
 			secondPositionUpperPrice: DefaultUpperPrice,
@@ -1612,7 +1612,7 @@ var (
 		"spread factor 3: two positions with consecutive price ranges: usdc (in) -> eth (out) (0.1% spread factor) | ofz": {
 			tokenOut:                 sdk.NewCoin(ETH, osmomath.NewInt(1820630)),
 			tokenInDenom:             USDC,
-			priceLimit:               osmomath.NewDec(6106),
+			priceLimit:               osmomath.NewBigDec(6106),
 			spreadFactor:             osmomath.MustNewDecFromStr("0.001"),
 			secondPositionLowerPrice: osmomath.NewDec(5500), // 315000
 			secondPositionUpperPrice: osmomath.NewDec(6250), // 322500
@@ -1671,7 +1671,7 @@ var (
 		"spread factor 4: two positions with partially overlapping price ranges: eth (in) -> usdc (out) (10% spread factor) | zfo": {
 			tokenOut:                 sdk.NewCoin(USDC, osmomath.NewInt(9321276930)),
 			tokenInDenom:             ETH,
-			priceLimit:               osmomath.NewDec(4128),
+			priceLimit:               osmomath.NewBigDec(4128),
 			spreadFactor:             osmomath.MustNewDecFromStr("0.1"),
 			secondPositionLowerPrice: osmomath.NewDec(4000),
 			secondPositionUpperPrice: osmomath.NewDec(4999),
@@ -1743,7 +1743,7 @@ var (
 		"spread factor 5: two positions with partially overlapping price ranges, not utilizing full liquidity of second position: usdc (in) -> eth (out) (5% spread factor) | ofz": {
 			tokenOut:                 sdk.NewCoin(ETH, osmomath.NewInt(1609138)),
 			tokenInDenom:             USDC,
-			priceLimit:               osmomath.NewDec(6056),
+			priceLimit:               osmomath.NewBigDec(6056),
 			spreadFactor:             osmomath.MustNewDecFromStr("0.05"),
 			secondPositionLowerPrice: osmomath.NewDec(5001),
 			secondPositionUpperPrice: osmomath.NewDec(6250),
@@ -1811,7 +1811,7 @@ var (
 		"spread factor 6: two sequential positions with a gap usdc (in) -> eth (out) (0.03% spread factor) | ofz": {
 			tokenOut:                 sdk.NewCoin(ETH, osmomath.NewInt(1820545)),
 			tokenInDenom:             USDC,
-			priceLimit:               osmomath.NewDec(6106),
+			priceLimit:               osmomath.NewBigDec(6106),
 			spreadFactor:             osmomath.MustNewDecFromStr("0.0003"),
 			secondPositionLowerPrice: osmomath.NewDec(5501), // 315010
 			secondPositionUpperPrice: osmomath.NewDec(6250), // 322500
@@ -1869,7 +1869,7 @@ var (
 		"spread factor 7: single position within one tick, trade completes but slippage protection interrupts trade early: usdc (in) -> eth (out) (1% spread factor) | ofz": {
 			tokenOut:     sdk.NewCoin(ETH, osmomath.NewInt(1820545)),
 			tokenInDenom: USDC,
-			priceLimit:   osmomath.NewDec(5002),
+			priceLimit:   osmomath.NewBigDec(5002),
 			spreadFactor: osmomath.MustNewDecFromStr("0.01"),
 			// from math import *
 			// from decimal import *
@@ -1902,14 +1902,14 @@ var (
 		"single position within one tick, trade does not complete due to lack of liquidity: usdc -> eth ": {
 			tokenOut:     sdk.NewCoin("usdc", osmomath.NewInt(5300000000)),
 			tokenInDenom: "eth",
-			priceLimit:   osmomath.NewDec(6000),
+			priceLimit:   osmomath.NewBigDec(6000),
 			spreadFactor: osmomath.ZeroDec(),
 			expectErr:    true,
 		},
 		"single position within one tick, trade does not complete due to lack of liquidity: eth -> usdc ": {
 			tokenOut:     sdk.NewCoin("eth", osmomath.NewInt(1100000)),
 			tokenInDenom: "usdc",
-			priceLimit:   osmomath.NewDec(4000),
+			priceLimit:   osmomath.NewBigDec(4000),
 			spreadFactor: osmomath.ZeroDec(),
 			expectErr:    true,
 		},
@@ -2115,7 +2115,7 @@ func (s *KeeperTestSuite) TestSwap_NoPositions() {
 	_, _, _, err := s.App.ConcentratedLiquidityKeeper.SwapInAmtGivenOut(
 		s.Ctx, s.TestAccs[0], pool,
 		DefaultCoin0, DefaultCoin1.Denom,
-		osmomath.ZeroDec(), osmomath.ZeroDec(),
+		osmomath.ZeroDec(), osmomath.ZeroBigDec(),
 	)
 	s.Require().Error(err)
 	s.Require().ErrorIs(err, types.NoSpotPriceWhenNoLiquidityError{PoolId: pool.GetId()})
@@ -2123,7 +2123,7 @@ func (s *KeeperTestSuite) TestSwap_NoPositions() {
 	_, _, _, err = s.App.ConcentratedLiquidityKeeper.SwapOutAmtGivenIn(
 		s.Ctx, s.TestAccs[0], pool,
 		DefaultCoin0, DefaultCoin1.Denom,
-		osmomath.ZeroDec(), osmomath.ZeroDec(),
+		osmomath.ZeroDec(), osmomath.ZeroBigDec(),
 	)
 
 	s.Require().Error(err)
@@ -2837,7 +2837,7 @@ func (s *KeeperTestSuite) TestInverseRelationshipSwapOutAmtGivenIn() {
 			secondTokenIn, secondTokenOut, _, err := s.App.ConcentratedLiquidityKeeper.SwapOutAmtGivenIn(
 				s.Ctx, s.TestAccs[0], poolBefore,
 				firstTokenOut, firstTokenIn.Denom,
-				DefaultZeroSpreadFactor, osmomath.ZeroDec(),
+				DefaultZeroSpreadFactor, osmomath.ZeroBigDec(),
 			)
 			s.Require().NoError(err)
 
@@ -2914,7 +2914,7 @@ func (s *KeeperTestSuite) TestInverseRelationshipSwapInAmtGivenOut() {
 			secondTokenIn, secondTokenOut, _, err := s.App.ConcentratedLiquidityKeeper.SwapInAmtGivenOut(
 				s.Ctx, s.TestAccs[0], poolBefore,
 				firstTokenIn, firstTokenOut.Denom,
-				DefaultZeroSpreadFactor, osmomath.ZeroDec(),
+				DefaultZeroSpreadFactor, osmomath.ZeroBigDec(),
 			)
 			s.Require().NoError(err)
 
@@ -3180,7 +3180,7 @@ func (s *KeeperTestSuite) TestFunctionalSwaps() {
 	//                              5000
 
 	// Swap multiple times USDC for ETH, therefore increasing the spot price
-	_, _, totalTokenIn, totalTokenOut := s.swapAndTrackXTimesInARow(clPool.GetId(), swapCoin1, ETH, types.MaxSpotPrice, positions.numSwaps)
+	_, _, totalTokenIn, totalTokenOut := s.swapAndTrackXTimesInARow(clPool.GetId(), swapCoin1, ETH, types.MaxSpotPriceBigDec, positions.numSwaps)
 	clPool, err := s.App.ConcentratedLiquidityKeeper.GetPoolById(s.Ctx, clPool.GetId())
 	s.Require().NoError(err)
 
@@ -3229,7 +3229,7 @@ func (s *KeeperTestSuite) TestFunctionalSwaps() {
 	}
 
 	// Swap multiple times ETH for USDC, therefore decreasing the spot price
-	_, _, totalTokenIn, totalTokenOut = s.swapAndTrackXTimesInARow(clPool.GetId(), swapCoin0, USDC, types.MinSpotPrice, positions.numSwaps)
+	_, _, totalTokenIn, totalTokenOut = s.swapAndTrackXTimesInARow(clPool.GetId(), swapCoin0, USDC, types.MinSpotPriceBigDec, positions.numSwaps)
 	clPool, err = s.App.ConcentratedLiquidityKeeper.GetPoolById(s.Ctx, clPool.GetId())
 	s.Require().NoError(err)
 
@@ -3289,7 +3289,7 @@ func (s *KeeperTestSuite) TestFunctionalSwaps() {
 	}
 
 	// Swap multiple times USDC for ETH, therefore increasing the spot price
-	_, _, totalTokenIn, totalTokenOut = s.swapAndTrackXTimesInARow(clPool.GetId(), swapCoin1, ETH, types.MaxSpotPrice, positions.numSwaps)
+	_, _, totalTokenIn, totalTokenOut = s.swapAndTrackXTimesInARow(clPool.GetId(), swapCoin1, ETH, types.MaxSpotPriceBigDec, positions.numSwaps)
 	clPool, err = s.App.ConcentratedLiquidityKeeper.GetPoolById(s.Ctx, clPool.GetId())
 	s.Require().NoError(err)
 
@@ -3349,7 +3349,7 @@ func (s *KeeperTestSuite) TestFunctionalSwaps() {
 	}
 
 	// Swap multiple times ETH for USDC, therefore decreasing the spot price
-	_, _, totalTokenIn, totalTokenOut = s.swapAndTrackXTimesInARow(clPool.GetId(), swapCoin0, USDC, types.MinSpotPrice, positions.numSwaps)
+	_, _, totalTokenIn, totalTokenOut = s.swapAndTrackXTimesInARow(clPool.GetId(), swapCoin0, USDC, types.MinSpotPriceBigDec, positions.numSwaps)
 	clPool, err = s.App.ConcentratedLiquidityKeeper.GetPoolById(s.Ctx, clPool.GetId())
 	s.Require().NoError(err)
 
@@ -3404,11 +3404,11 @@ func (s *KeeperTestSuite) TestInfiniteSwapLoop_OutGivenIn() {
 	swapEthFunded := sdk.NewCoin(ETH, osmomath.Int(osmomath.MustNewDecFromStr("10000000000000000000000000000000000000000")))
 	swapUSDCFunded := sdk.NewCoin(USDC, osmomath.Int(osmomath.MustNewDecFromStr("10000")))
 	s.FundAcc(swapAddress, sdk.NewCoins(swapEthFunded, swapUSDCFunded))
-	_, tokenOut, _, err := s.clk.SwapInAmtGivenOut(s.Ctx, swapAddress, pool, sdk.NewCoin(USDC, osmomath.NewInt(10000)), ETH, pool.GetSpreadFactor(s.Ctx), osmomath.ZeroDec())
+	_, tokenOut, _, err := s.clk.SwapInAmtGivenOut(s.Ctx, swapAddress, pool, sdk.NewCoin(USDC, osmomath.NewInt(10000)), ETH, pool.GetSpreadFactor(s.Ctx), osmomath.ZeroBigDec())
 	s.Require().NoError(err)
 
 	// Swap back in the amount that was swapped out to test the inverse relationship
-	_, _, _, err = s.clk.SwapOutAmtGivenIn(s.Ctx, swapAddress, pool, tokenOut, ETH, pool.GetSpreadFactor(s.Ctx), osmomath.ZeroDec())
+	_, _, _, err = s.clk.SwapOutAmtGivenIn(s.Ctx, swapAddress, pool, tokenOut, ETH, pool.GetSpreadFactor(s.Ctx), osmomath.ZeroBigDec())
 	s.Require().NoError(err)
 }